Add instructions on releasing "latest"
My infrastructure code is configured to just deploy the latest version
of the image that I've shipped to my private registry, which is why I
need a latest tag.

I do want to be able to check which version that is though, so need to
tag it with a version number too. This approach seems like a nice and
lightweight way of handling that, without requiring that every image
that's built gets tagged as latest. And let's face it, unless we are
building the latest version, tagging it as "latest" isn't necessarily
going to be appropriate.